deps(composer): Add deployer/deployer dependency
The "deployer/deployer" package with version "^7.5" was added to the composer.json. This addition is accompanied by changes in composer.lock and the creation of a new deploy.yaml file. The purpose is likely to set up a deployment process for the project, as the deployer package is a deployment tool.